messaging/error updates

+++ src/@providers/terraform.service.ts
@@ -83,9 +83,6 @@ export abstract class TerraformResourceService<
     );
 
     const status = await cmd.status;
-    if (!status.success) {
-      throw new Error(`Terraform init failed with exit code ${status.code}`);
-    }
 
     return {
       ...status,
@@ -123,9 +120,6 @@ export abstract class TerraformResourceService<
     );
 
     const status = await cmd.status;
-    if (!status.success) {
-      throw new Error(`Terraform plan failed with exit code ${status.code}`);
-    }
 
     return {
       ...status,
@@ -163,9 +157,6 @@ export abstract class TerraformResourceService<
     );
 
     const status = await cmd.status;
-    if (!status.success) {
-      throw new Error(`Terraform apply failed with exit code ${status.code}`);
-    }
 
     return {
       ...status,
@@ -204,9 +195,6 @@ export abstract class TerraformResourceService<
     );
 
     const status = await cmd.status;
-    if (!status.success) {
-      throw new Error(`Terraform output failed with exit code ${status.code}`);
-    }
 
     return {
       ...status,
@@ -261,7 +249,11 @@ export abstract class TerraformResourceService<
       const imports = await module.genImports(options.state.id);
 
       // We have to run this before we can run `terraform import`
-      await this.tfInit(cwd, stack, options.logger);
+      const { stderr: init_stderr } = await this.tfInit(cwd, stack, options.logger);
+      if (init_stderr && init_stderr.length > 0) {
+        subscriber.error(new TextDecoder().decode(init_stderr));
+        return;
+      }
       initRan = true;
 
       const terraform = await this.getTerraformPlugin();
@@ -279,7 +271,11 @@ export abstract class TerraformResourceService<
         },
       });
 
-      await this.tfInit(cwd, stack, options.logger);
+      const { stderr: init_stderr } = await this.tfInit(cwd, stack, options.logger);
+      if (init_stderr && init_stderr.length > 0) {
+        subscriber.error(new TextDecoder().decode(init_stderr));
+        return;
+      }
     }
 
     subscriber.next({
@@ -290,7 +286,11 @@ export abstract class TerraformResourceService<
       },
     });
 
-    await this.tfPlan(cwd, options.logger);
+    const { stderr: plan_stderr } = await this.tfPlan(cwd, options.logger);
+    if (plan_stderr && plan_stderr.length > 0) {
+      subscriber.error(new TextDecoder().decode(plan_stderr));
+      return;
+    }
 
     subscriber.next({
       status: {
@@ -300,9 +300,9 @@ export abstract class TerraformResourceService<
       },
     });
 
-    const { stderr } = await this.tfApply(cwd, options.logger);
-    if (stderr && stderr.length > 0) {
-      subscriber.error(new TextDecoder().decode(stderr));
+    const { stderr: apply_stderr } = await this.tfApply(cwd, options.logger);
+    if (apply_stderr && apply_stderr.length > 0) {
+      subscriber.error(new TextDecoder().decode(apply_stderr));
       return;
     }
 
@@ -323,7 +323,11 @@ export abstract class TerraformResourceService<
       lockFile: new TextDecoder().decode(lockFileBuffer),
     };
 
-    const { stdout: rawOutputs } = await this.tfOutput(cwd, options.logger);
+    const { stdout: rawOutputs, stderr: output_stderr } = await this.tfOutput(cwd, options.logger);
+    if (output_stderr && output_stderr.length > 0) {
+      subscriber.error(new TextDecoder().decode(output_stderr));
+      return;
+    }
     const parsedOutputs = JSON.parse(new TextDecoder().decode(rawOutputs));
 
     await Deno.remove(cwd, { recursive: true });
@@ -418,7 +422,11 @@ export abstract class TerraformResourceService<
         },
       });
 
-      await this.tfInit(cwd, stack, options.logger);
+      const { stderr: init_stderr } = await this.tfInit(cwd, stack, options.logger);
+      if (init_stderr && init_stderr.length > 0) {
+        subscriber.error(new TextDecoder().decode(init_stderr));
+        return;
+      }
 
       subscriber.next({
         status: {
@@ -428,7 +436,11 @@ export abstract class TerraformResourceService<
         },
       });
 
-      await this.tfPlan(cwd, options.logger);
+      const { stderr: plan_stderr } = await this.tfPlan(cwd, options.logger);
+      if (plan_stderr && plan_stderr.length > 0) {
+        subscriber.error(new TextDecoder().decode(plan_stderr));
+        return;
+      }
 
       subscriber.next({
         status: {
@@ -438,9 +450,9 @@ export abstract class TerraformResourceService<
         },
       });
 
-      const { stderr } = await this.tfApply(options.cwd, options.logger);
-      if (stderr && stderr.length > 0) {
-        subscriber.error(new TextDecoder().decode(stderr));
+      const { stderr: apply_stderr } = await this.tfApply(options.cwd, options.logger);
+      if (apply_stderr && apply_stderr.length > 0) {
+        subscriber.error(new TextDecoder().decode(apply_stderr));
         return;
       }
